feat(scripts): change caps-to-foundation to transfer-ownership
